refactor BranchList
